+++ CHANGELOG.md
@@ -6,6 +6,7 @@ BUG FIXES:
 
 IMPROVEMENTS:
 * cli: Add flags to configure Nomad API client [[GH-213](https://github.com/hashicorp/nomad-pack/pull/213)]
+* template: Add support for custom Spew configurations. [[GH-220](https://github.com/hashicorp/nomad-pack/pull/220)]
 
 ## 0.0.1-techpreview2 (February 07, 2022)